+++ packages/std/src/addresses.rs
@@ -9,7 +9,7 @@ use std::fmt;
 use std::ops::Deref;
 use thiserror::Error;
 
-use crate::{binary::Binary, HexBinary};
+use crate::{binary::Binary, forward_ref_partial_eq, HexBinary};
 
 /// A human readable address.
 ///
@@ -31,6 +31,8 @@ use crate::{binary::Binary, HexBinary};
     Serialize, Deserialize, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, JsonSchema,
 )]
 pub struct Addr(String);
+
+forward_ref_partial_eq!(Addr, Addr);
 
 impl Addr {
     /// Creates a new `Addr` instance from the given input without checking the validity
@@ -86,6 +88,9 @@ impl AsRef<str> for Addr {
 }
 
 /// Implement `Addr == &str`
+///
+/// Deprecated. This comparison unsafe. Convert both sides to Addr first.
+/// Will be removed soon: https://github.com/CosmWasm/cosmwasm/issues/1669
 impl PartialEq<&str> for Addr {
     fn eq(&self, rhs: &&str) -> bool {
         self.0 == *rhs
@@ -93,6 +98,9 @@ impl PartialEq<&str> for Addr {
 }
 
 /// Implement `&str == Addr`
+///
+/// Deprecated. This comparison unsafe. Convert both sides to Addr first.
+/// Will be removed soon: https://github.com/CosmWasm/cosmwasm/issues/1669
 impl PartialEq<Addr> for &str {
     fn eq(&self, rhs: &Addr) -> bool {
         *self == rhs.0
@@ -100,6 +108,9 @@ impl PartialEq<Addr> for &str {
 }
 
 /// Implement `Addr == String`
+///
+/// Deprecated. This comparison unsafe. Convert both sides to Addr first.
+/// Will be removed soon: https://github.com/CosmWasm/cosmwasm/issues/1669
 impl PartialEq<String> for Addr {
     fn eq(&self, rhs: &String) -> bool {
         &self.0 == rhs
@@ -107,6 +118,9 @@ impl PartialEq<String> for Addr {
 }
 
 /// Implement `String == Addr`
+///
+/// Deprecated. This comparison unsafe. Convert both sides to Addr first.
+/// Will be removed soon: https://github.com/CosmWasm/cosmwasm/issues/1669
 impl PartialEq<Addr> for String {
     fn eq(&self, rhs: &Addr) -> bool {
         self == &rhs.0
@@ -429,24 +443,34 @@ mod tests {
         assert_eq!(addr.as_ref(), "literal-string");
     }
 
+    // Please note that this will be removed soon
+    // https://github.com/CosmWasm/cosmwasm/issues/1669
     #[test]
-    fn addr_implements_partial_eq_with_str() {
+    fn addr_implements_partial_eq_with_str_and_string() {
         let addr = Addr::unchecked("cos934gh9034hg04g0h134");
 
         // `Addr == &str`
         assert_eq!(addr, "cos934gh9034hg04g0h134");
         // `&str == Addr`
         assert_eq!("cos934gh9034hg04g0h134", addr);
-    }
-
-    #[test]
-    fn addr_implements_partial_eq_with_string() {
-        let addr = Addr::unchecked("cos934gh9034hg04g0h134");
-
         // `Addr == String`
         assert_eq!(addr, String::from("cos934gh9034hg04g0h134"));
         // `String == Addr`
         assert_eq!(String::from("cos934gh9034hg04g0h134"), addr);
+    }
+
+    #[test]
+    fn addr_implements_partial_eq_addr_ref() {
+        let addr = Addr::unchecked("cos934gh9034hg04g0h134");
+        let addr_ref = &addr;
+        let addr_ref2 = &addr;
+
+        // `Addr == &Addr`
+        assert_eq!(addr, addr_ref);
+        // `&Addr == Addr`
+        assert_eq!(addr_ref, addr);
+        // `&Addr == &Addr`
+        assert_eq!(addr_ref, addr_ref2);
     }
 
     #[test]