+++ packages/std/src/math/uint128.rs
@@ -12,7 +12,9 @@ use crate::errors::{
     CheckedMultiplyFractionError, CheckedMultiplyRatioError, DivideByZeroError, OverflowError,
     OverflowOperation, StdError,
 };
-use crate::{impl_mul_fraction, ConversionOverflowError, Fraction, Uint256, Uint64};
+use crate::{
+    forward_ref_partial_eq, impl_mul_fraction, ConversionOverflowError, Fraction, Uint256, Uint64,
+};
 
 /// A thin wrapper around u128 that is using strings for JSON encoding/decoding,
 /// such that the full u128 range can be used for clients that convert JSON numbers to floats,
@@ -35,6 +37,8 @@ use crate::{impl_mul_fraction, ConversionOverflowError, Fraction, Uint256, Uint6
 /// ```
 #[derive(Copy, Clone, Default, Debug, PartialEq, Eq, PartialOrd, Ord, JsonSchema)]
 pub struct Uint128(#[schemars(with = "String")] u128);
+
+forward_ref_partial_eq!(Uint128, Uint128);
 
 impl Uint128 {
     pub const MAX: Self = Self(u128::MAX);
@@ -65,19 +69,23 @@ impl Uint128 {
     }
 
     /// Returns a copy of the number as big endian bytes.
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     pub const fn to_be_bytes(self) -> [u8; 16] {
         self.0.to_be_bytes()
     }
 
     /// Returns a copy of the number as little endian bytes.
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     pub const fn to_le_bytes(self) -> [u8; 16] {
         self.0.to_le_bytes()
     }
 
+    #[must_use]
     pub const fn is_zero(&self) -> bool {
         self.0 == 0
     }
 
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     pub fn pow(self, exp: u32) -> Self {
         self.0.pow(exp).into()
     }
@@ -86,6 +94,7 @@ impl Uint128 {
     ///
     /// Due to the nature of the integer division involved, the result is always floored.
     /// E.g. 5 * 99/100 = 4.
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     pub fn multiply_ratio<A: Into<u128>, B: Into<u128>>(
         &self,
         numerator: A,
@@ -132,6 +141,7 @@ impl Uint128 {
     /// let result = a.full_mul(2u32);
     /// assert_eq!(result.to_string(), "680564733841876926926749214863536422910");
     /// ```
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     pub fn full_mul(self, rhs: impl Into<u128>) -> Uint256 {
         Uint256::from(self.u128())
             .checked_mul(Uint256::from(rhs.into()))
@@ -187,42 +197,51 @@ impl Uint128 {
             .ok_or_else(|| DivideByZeroError::new(self))
     }
 
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     #[inline]
     pub fn wrapping_add(self, other: Self) -> Self {
         Self(self.0.wrapping_add(other.0))
     }
 
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     #[inline]
     pub fn wrapping_sub(self, other: Self) -> Self {
         Self(self.0.wrapping_sub(other.0))
     }
 
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     #[inline]
     pub fn wrapping_mul(self, other: Self) -> Self {
         Self(self.0.wrapping_mul(other.0))
     }
 
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     #[inline]
     pub fn wrapping_pow(self, other: u32) -> Self {
         Self(self.0.wrapping_pow(other))
     }
 
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     pub fn saturating_add(self, other: Self) -> Self {
         Self(self.0.saturating_add(other.0))
     }
 
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     pub fn saturating_sub(self, other: Self) -> Self {
         Self(self.0.saturating_sub(other.0))
     }
 
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     pub fn saturating_mul(self, other: Self) -> Self {
         Self(self.0.saturating_mul(other.0))
     }
 
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     pub fn saturating_pow(self, exp: u32) -> Self {
         Self(self.0.saturating_pow(exp))
     }
 
+    #[must_use = "this returns the result of the operation, without modifying the original"]
     pub const fn abs_diff(self, other: Self) -> Self {
         Self(if self.0 < other.0 {
             other.0 - self.0
@@ -527,18 +546,6 @@ where
     }
 }
 
-impl PartialEq<&Uint128> for Uint128 {
-    fn eq(&self, rhs: &&Uint128) -> bool {
-        self == *rhs
-    }
-}
-
-impl PartialEq<Uint128> for &Uint128 {
-    fn eq(&self, rhs: &Uint128) -> bool {
-        *self == rhs
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use crate::errors::CheckedMultiplyFractionError::{ConversionOverflow, DivideByZero};
@@ -821,7 +828,7 @@ mod tests {
     #[test]
     #[should_panic]
     fn uint128_pow_overflow_panics() {
-        Uint128::MAX.pow(2u32);
+        _ = Uint128::MAX.pow(2u32);
     }
 
     #[test]
@@ -867,7 +874,7 @@ mod tests {
     #[test]
     #[should_panic(expected = "Denominator must not be zero")]
     fn uint128_multiply_ratio_panics_for_zero_denominator() {
-        Uint128(500).multiply_ratio(1u128, 0u128);
+        _ = Uint128(500).multiply_ratio(1u128, 0u128);
     }
 
     #[test]
@@ -1095,7 +1102,7 @@ mod tests {
     #[should_panic(expected = "ConversionOverflowError")]
     fn mul_floor_panics_on_overflow() {
         let fraction = (21u128, 8u128);
-        Uint128::MAX.mul_floor(fraction);
+        _ = Uint128::MAX.mul_floor(fraction);
     }
 
     #[test]
@@ -1115,7 +1122,7 @@ mod tests {
     #[should_panic(expected = "DivideByZeroError")]
     fn mul_floor_panics_on_zero_div() {
         let fraction = (21u128, 0u128);
-        Uint128::new(123456).mul_floor(fraction);
+        _ = Uint128::new(123456).mul_floor(fraction);
     }
 
     #[test]
@@ -1178,7 +1185,7 @@ mod tests {
     #[should_panic(expected = "ConversionOverflowError")]
     fn mul_ceil_panics_on_overflow() {
         let fraction = (21u128, 8u128);
-        Uint128::MAX.mul_ceil(fraction);
+        _ = Uint128::MAX.mul_ceil(fraction);
     }
 
     #[test]
@@ -1198,7 +1205,7 @@ mod tests {
     #[should_panic(expected = "DivideByZeroError")]
     fn mul_ceil_panics_on_zero_div() {
         let fraction = (21u128, 0u128);
-        Uint128::new(123456).mul_ceil(fraction);
+        _ = Uint128::new(123456).mul_ceil(fraction);
     }
 
     #[test]
@@ -1216,7 +1223,7 @@ mod tests {
     #[should_panic(expected = "DivideByZeroError")]
     fn div_floor_raises_with_zero() {
         let fraction = (Uint128::zero(), Uint128::new(21));
-        Uint128::new(123456).div_floor(fraction);
+        _ = Uint128::new(123456).div_floor(fraction);
     }
 
     #[test]
@@ -1267,7 +1274,7 @@ mod tests {
     #[should_panic(expected = "ConversionOverflowError")]
     fn div_floor_panics_on_overflow() {
         let fraction = (8u128, 21u128);
-        Uint128::MAX.div_floor(fraction);
+        _ = Uint128::MAX.div_floor(fraction);
     }
 
     #[test]
@@ -1287,7 +1294,7 @@ mod tests {
     #[should_panic(expected = "DivideByZeroError")]
     fn div_ceil_raises_with_zero() {
         let fraction = (Uint128::zero(), Uint128::new(21));
-        Uint128::new(123456).div_ceil(fraction);
+        _ = Uint128::new(123456).div_ceil(fraction);
     }
 
     #[test]
@@ -1338,7 +1345,7 @@ mod tests {
     #[should_panic(expected = "ConversionOverflowError")]
     fn div_ceil_panics_on_overflow() {
         let fraction = (8u128, 21u128);
-        Uint128::MAX.div_ceil(fraction);
+        _ = Uint128::MAX.div_ceil(fraction);
     }
 
     #[test]